Fix two defects and add logging
There were a couple of defects that were quite bad for users, but also quite
bad in their origins. One was that the web view wasn't removed from the view
hierarchy, which could lead to an unresponsive white box being displayed
instead of the zooming user interface. The other was that the setting of the
web view frame wasn't repeated when the device orientation changed.

Really bad hurdle in the way of fixing those defects was that there's no
possibility to connect the debugger, even for logs. So, there's now a few
methods to get visible logs in the keyboard UI itself. Code changes are
required to activate the features, so the methods are for developers only, at
this time.

Removed fixed defects from readme. Added new items to readme.