Finalise first integration with UITextChecker
This is based on the prototype provided from the lblasa-master branch, but modified as follows.

-   The get() and get...weights() methods use async and await instead of
explicit Promise code. Also weights are returned as a JS Map.

-   The Swift end of the predict command expects a sentence and offers
replacements for the last word. The JS end now doesn't cache, because that
doesn't seem worthwhile any more. The PredictorCompletions class is a subclass
of Predictor and overrides the get...weights() method.

-   Added a `contraction` group with apostrophe and hyphen characters, because
these can be present in UITextChecker predictions.

-   The top-level predict command handler throws on error instead of returning
an empty dictionary. If the ZoomBox ready throws, then the error is re-thrown
and set into the message.

-   There's some code to experiment with UITextChecker behaviour. The results
can only be viewed in the debugger so it's only for developers at this point.