Add predictions to custom keyboard for iOS
The dummy selection is now functional.

Moved Swift end of the predictions to a common code directory and group.

Removed some diagnostic logging.

Changed the interface to the JS PredictorCompletions to being a bound send
command instead of the bridge object. Commands from the predictor now get
routed via the layer above the zooming UI and can be transcribed in the log.