Add "tails" of smaller characters like original Dasher and some backlog items
Tail characters appear to the right of where the smallest boxes spawn. This is
governed by the solver configuration. A white rectangle has been added to the
user interface to hide the right hand end of zoom boxes that have been rendered.
Spawing rules were rationalised and simplified as part of this change. Also,
some properties were moved from the ZoomBox to the Limits.

An underlying problem in move application was fixed. A move now won't be
applied to a box if it would cause the box to leave the solvable area.

Adding tail characters required an extra right shifting of character position,
to prevent them being rendered underneath the characters in parent boxes.

There was a problem that a character in the zooming user interface could be
partially covered by a child box, if the child's visual weight was high enough.
The View structure now has two layers of SVG so that even the highest
rendered rectangle is below the lowest rendered character. Obliquely, the
hiding white rectangle sits between the two layers, so it hides rectangles but
doesn't hide characters.

There were some changes to the order in which HTML and SVG items get created
in UserInterface.

Added a couple of features for developers:
-   Random Predictor option.
-   Frozen mode.