Safety commit for dark mode support and notification icon customisation, still with some unused and commented out code.
Added an informative feature to the base application UI and inherited it to all
others. The text in the UI can now be tapped to cycle through informative
messages for: integration level, device dark mode selection, and build flavour.

Added notification icon customisation. Added a UI feature to show a notification
on demand. This is to demonstrate the Android capability and type of small icon
that is supported. Increased the minSdkVersion to 26 to facilitate notification
demonstration. Added a one-colour brand logo that is suitable for use as a
notification icon.

Changed the base application to a theme that supports dark mode.