Adding the data portion of Enron mobile dataset originally developed by Keith Vertanen:
    https://www.keithv.com/software/enronmobile/

Original synopsis:
------------------
This dataset consists of sentences written by Enron employees on BlackBerry mobile devices. We provide a series of test sets we recommend for use in text entry evaluations.

The sentence and sentence fragments were found by looking for messages with the default BlackBerry signature at the end of an email. All the sentences were manually reviewed and corrected. For each sentence, we provide metadata about the category (business, personal, Enron-specific), how easy the sentence was to remember, and how quickly and accurately the sentence was typed on full-sized keyboards.