Move the select list builder into a separate method and reinstate deep copying in the option list setter
The select list builder code was previously copied in two places. Now that it's
more code it seems justifiable to move it to a function.
The optionStrings setter had a crafty .slice() call that actually would have
deep copied the strings. Now that it's optionList the deep copy has to be made
more manual and explicit.
Also, added a constant for typeof string.

Update copyright notices.

Signed-off-by: Jim Hawkins <sjjhsjjhsjjh@gmail.com>