Fix the voice selection back end
The voice selection in the UserInterface relied on the voices in the selection
drop-down being in the same order as in the Speech object. I think it wasn't
reliable either. Anyway, now that voices are grouped they aren't guarranteed to
be in the same orders in the Speech object and the selection drop-down.

Fixed that by finding the name whenever a selection is made or the speak-on-stop
box is ticked.

Signed-off-by: Jim Hawkins <sjjhsjjhsjjh@gmail.com>